backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
email-validator==2.1.0
python-dotenv==1.0.0
pyrogram==2.0.106
tgcrypto==1.2.5
celery==5.3.4
redis==5.0.1
google-generativeai==0.3.1
websockets==12.0
aiofiles==23.2.1
httpx==0.25.2
Pillow==10.1.0

